backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt
better-exceptions==0.2.2
python-multipart==0.0.6
alembic==1.13.1
asyncpg==0.29.0
email-validator==2.3.0
alembic==1.13.1
psycopg2-binary
openai>=1.0.0